+++ CHANGES.md
@@ -2,6 +2,14 @@ ch.vorburger.xtendbeans Release Notes
 =====================================
 
 
+v1.1.0 @ TBD
+---
+
+* AssertBeans without Object.equals()
+* Builder class to use can be customized (useful if it cannot be auto-detected)
+* Documentation in README
+
+
 v1.0.0 @ 2016-08-10
 ---